requirements.txt
langchain
langchain-groq
langchain-community
langgraph
duckduckgo-search
streamlit
python-dotenv